+++ app/Models/Payment.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+
 class Payment extends Model
 {
     use HasFactory;
@@ -21,55 +22,19 @@ class Payment extends Model
         'payment_method',
     ];
 
-
-
-     public function livre()
+    /**
+     * Relation avec le livre
+     */
+    public function livre()
     {
         return $this->belongsTo(Livre::class, 'livre_id');
     }
 
+    /**
+     * Relation avec l’utilisateur
+     */
     public function user()
     {
         return $this->belongsTo(User::class, 'user_id');
-
-/*use Illuminate\Database\Eloquent\Model;
-
-class Payment extends Model
-{
-    protected $fillable = [
-        'user_id',
-        'subscription_id',
-        'amount',
-        'payment_method',
-        'transaction_id',
-        'status',
-        'payment_data'
-    ];
-
-    protected $casts = [
-        'payment_data' => 'array',
-        'amount' => 'decimal:2'
-    ];
-
-    public function user()
-    {
-        return $this->belongsTo(User::class);
     }
-
-    public function subscription()
-    {
-        return $this->belongsTo(Subscription::class);
-    }
-
-    public function isCompleted()
-    {
-        return $this->status === 'completed';
-    }
-
-    public function isFailed()
-    {
-        return $this->status === 'failed';
-
-    }
-  */
 }
